Move collection and principal query properties
